Finish responding to review comments; continue fixing tests
